requirements.txt
discord.py==1.6.0
Pillow==7.0.0
aiohttp==3.7.3
googletrans==3.0.0
py3-pinterest==1.2.2
PyNaCl==1.4.0
youtube-dl==2021.1.24.1